Make site loading pixels to be shared
