+++ iOS/DuckDuckGo/TabViewController.swift
@@ -1470,6 +1470,11 @@ class TabViewController: UIViewController {
         }
     }
 
+    /// Set by `webView(_:decidePolicyFor:decisionHandler:)` for main-frame navigations and consumed by
+    /// `webView(_:didStartProvisionalNavigation:)`, which stamps it onto the resulting `WKNavigation` for
+    /// later use by the site-loading success/failure pixels.
+    var pendingSiteLoadingNavigationType: String?
+
     private func resetDashboardInfo() {
         if let url = url {
             if didGoBackForward, let privacyInfo = previousPrivacyInfosByURL[url] {
@@ -1904,6 +1909,7 @@ extension TabViewController: WKNavigationDelegate {
     }
 
     func webView(_ webView: WKWebView, didStartProvisionalNavigation navigation: WKNavigation!) {
+        recordSiteLoadingStart(for: navigation)
         lastError = nil
         lastRenderedURL = webView.url
         cancelTrackerNetworksAnimation()
@@ -1918,6 +1924,7 @@ extension TabViewController: WKNavigationDelegate {
     }
 
     func webView(_ webView: WKWebView, didFinish navigation: WKNavigation!) {
+        fireSiteLoadingSuccessPixel(for: navigation)
         self.preventUniversalLinksOnce = false
         self.currentlyLoadedURL = webView.url
         didFinishURLSubject.send(webView.url)
@@ -2252,6 +2259,7 @@ extension TabViewController: WKNavigationDelegate {
     }
 
     func webView(_ webView: WKWebView, didFail navigation: WKNavigation!, withError error: Error) {
+        fireSiteLoadingFailurePixel(for: navigation, error: error)
         Logger.general.debug("didFailNavigation; error: \(error)")
         adClickAttributionDetection.onDidFailNavigation()
         adClickExternalOpenDetector.failNavigation(error: error)
@@ -2280,6 +2288,7 @@ extension TabViewController: WKNavigationDelegate {
     }
     
     func webView(_ webView: WKWebView, didFailProvisionalNavigation navigation: WKNavigation!, withError error: Error) {
+        fireSiteLoadingFailurePixel(for: navigation, error: error)
         Logger.general.debug("didFailProvisionalNavigation; error: \(error)")
         adClickAttributionDetection.onDidFailNavigation()
         adClickExternalOpenDetector.failNavigation(error: error)
@@ -2369,6 +2378,8 @@ extension TabViewController: WKNavigationDelegate {
     func webView(_ webView: WKWebView,
                  decidePolicyFor navigationAction: WKNavigationAction,
                  decisionHandler: @escaping (WKNavigationActionPolicy) -> Void) {
+
+        captureSiteLoadingNavigationType(navigationAction)
 
         if let url = navigationAction.request.url {
             if !tabURLInterceptor.allowsNavigatingTo(url: url) {
